update prediction: DFHiC, iEnhance

+++ data_download_downsample.sh
@@ -4,9 +4,7 @@ set -u  # Treat unset variables as an error
 seed=42
 
 # examples
-# bash data_download_downsample.sh https://ftp.ncbi.nlm.nih.gov/geo/samples/GSM1551nnn/GSM1551550/suppl/GSM1551550_HIC001_merged_nodups.txt.gz GSM1551550_HIC001 hg19 5000000 ./juicer_tools.jar KR 10000 data
-# bash data_download_downsample.sh https://ftp.ncbi.nlm.nih.gov/geo/samples/GSM1551nnn/GSM1551620/suppl/GSM1551620_HIC071_merged_nodups.txt.gz GSM1551620_HIC071 hg19 5000000 ./juicer_tools.jar KR 10000 data_K562
-# bash data_download_downsample.sh https://ftp.ncbi.nlm.nih.gov/geo/samples/GSM1551nnn/GSM1551643/suppl/GSM1551643_HIC098_merged_nodups.txt.gz GSM1551643_HIC098 mm9.chrom.sizes 5000000 ./juicer_tools.jar KR 10000 data_ch12-lx
+# bash data_download_downsample.sh https://ftp.ncbi.nlm.nih.gov/geo/samples/GSM1551nnn/GSM1551550/suppl/GSM1551550_HIC001_merged_nodups.txt.gz GSM1551550_HIC001 hg19 5000000 ./juicer_tools.jar KR 10000 data_GM12878
 
 # 필요한 매개변수 체크
 if [ "$#" -ne 8 ]; then
@@ -14,7 +12,7 @@ if [ "$#" -ne 8 ]; then
   exit 1
 fi
 
-### you can modifiy here as string type with ""!!!
+### you can modify here as string type with ""!!!
 path=$(pwd)
 download_url=$1
 file_name=$2
@@ -29,7 +27,7 @@ echo ""
 echo "  ...current working directory is $(pwd)"
 echo "      data download url : ${download_url}"
 echo "      raw data file : ${file_name}.txt.gz"
-echo "      refrence genome : ${ref_genome}"
+echo "      reference genome : ${ref_genome}"
 echo "      downsampled read : ${downsampled_read}"
 echo "      juicertools : ${juicertools}" # path to juicer_tool.jar of docker image
 echo "      normalization method : ${normalization}"
@@ -47,37 +45,43 @@ wget "${download_url}" -O "${save_name}"
 # 텍스트 파일 경로 설정
 file_path=${path}/${save_name}
 
-# 저장할 샘플링 파일 경로 설정
-downsample_name=${path}/${file_name}_ds_${downsampled_read}.txt.gz
-
 # 파일의 총 라인 수 가져오기
 total_lines=$(zcat "$file_path" | wc -l)
 echo ""
 echo "  ...The number of total reads is ${total_lines}."
 
-# 중복 없이 무작위로 샘플링한 라인을 저장할 임시 파일 생성
-temp_file=$(mktemp)
+# downsampled_read 확인 및 조건 처리
+if [ "$total_lines" -le "$downsampled_read" ]; then
+  echo "  ...Total reads (${total_lines}) are less than or equal to the requested downsampled reads (${downsampled_read}). Skipping downsampling."
+  downsample_name="" # 이후에 관련 작업 건너뛰기 위해 빈 값 설정
+else
+  # 저장할 샘플링 파일 경로 설정
+  downsample_name=${path}/${file_name}_ds_${downsampled_read}.txt.gz
 
-# 원본 파일에서 중복 없이 무작위로 샘플링한 라인을 추출하여 임시 파일에 저장
-zcat "$file_path" | shuf -n "$sample_size" >> "$temp_file"
-echo "  ...And ${downsampled_read} reads are randomly sampled."
+  # 중복 없이 무작위로 샘플링한 라인을 저장할 임시 파일 생성
+  temp_file=$(mktemp)
 
-# 3번째와 7번째 컬럼(크로모좀) 기준으로 오름차순 정렬
-sort -k3,3d -k7,7 "$temp_file" | gzip > "${downsample_name}"
-rm "$temp_file"  # 임시 파일 삭제
-echo "  ...Randomly sampled reads are saved: ${downsample_name}"
+  # 원본 파일에서 중복 없이 무작위로 샘플링한 라인을 추출하여 임시 파일에 저장
+  zcat "$file_path" | shuf -n "$downsampled_read" >> "$temp_file"
+  echo "  ...And ${downsampled_read} reads are randomly sampled."
 
+  # 3번째와 7번째 컬럼(크로모좀) 기준으로 오름차순 정렬
+  sort -k3,3d -k7,7 "$temp_file" | gzip > "${downsample_name}"
+  rm "$temp_file"  # 임시 파일 삭제
+  echo "  ...Randomly sampled reads are saved: ${downsample_name}"
+fi
 
 ### .hic format ###
-# juicer pre 로 .hic 변환
 echo ""
 echo "  ...To hic,"
 java -Xmx2g -jar ${juicertools} pre ${save_name} ./${file_name}.hic ${ref_genome} # 원본 데이터
 echo "  ...${file_name}.hic is generated."
-echo ""
-java -Xmx2g -jar ${juicertools} pre ${downsample_name} ./${file_name}_ds_${downsampled_read}.hic ${ref_genome} # 다운 샘플링 데이터
-echo "  ...${file_name}_ds_${downsampled_read}.hic is generated."
 
+# downsample_name이 존재하는 경우에만 .hic 파일 생성
+if [ -n "$downsample_name" ]; then
+  java -Xmx2g -jar ${juicertools} pre ${downsample_name} ./${file_name}_ds_${downsampled_read}.hic ${ref_genome}
+  echo "  ...${file_name}_ds_${downsampled_read}.hic is generated."
+fi
 
 ### .txt format for each chromosome (intra) ###
 # ref_genome에 따라 크로모좀 범위 설정
@@ -92,8 +96,7 @@ else
 fi
 
 # 결과를 저장할 폴더 생성​​
-mkdir ${path}/${saved_in}
-# 전체 크로모좀에 대하여 반복문으로 명령 실행
+mkdir -p ${path}/${saved_in}
 for ((chrom=1; chrom<end; chrom++))
 do
     java -jar ${juicertools} dump observed ${normalization} ./${file_name}.hic ${chrom} ${chrom} BP ${resolution} ${path}/${saved_in}/chr${chrom}_${resolution}.txt
@@ -101,12 +104,13 @@ done
 echo ""
 echo "  ...Intra chromosome contact matrix are generated: ./${saved_in}/"
 
-# 결과를 저장할 폴더 생성​​
-mkdir ${path}/${saved_in}_downsampled_${downsampled_read}
-# 전체 크로모좀에 대하여 반복문으로 명령 실행
-for ((chrom=1; chrom<end; chrom++))
-do
-    java -jar ${juicertools} dump observed ${normalization} ./${file_name}_ds_${downsampled_read}.hic ${chrom} ${chrom} BP ${resolution} ${path}/${saved_in}_downsampled_${downsampled_read}/chr${chrom}_${resolution}.txt
-done
-echo "  ...Downsampled contact matrix are generated: ./${saved_in}_downsampled_${downsampled_read}/"
+# 다운샘플링 데이터에 대해서는 downsample_name이 존재하는 경우에만 실행
+if [ -n "$downsample_name" ]; then
+  mkdir -p ${path}/${saved_in}_downsampled_${downsampled_read}
+  for ((chrom=1; chrom<end; chrom++))
+  do
+      java -jar ${juicertools} dump observed ${normalization} ./${file_name}_ds_${downsampled_read}.hic ${chrom} ${chrom} BP ${resolution} ${path}/${saved_in}_downsampled_${downsampled_read}/chr${chrom}_${resolution}.txt
+  done
+  echo "  ...Downsampled contact matrix are generated: ./${saved_in}_downsampled_${downsampled_read}/"
+fi
 echo ""
